refactor finalize command
