+++ automation/Commands/browser_commands.py
@@ -467,17 +467,34 @@ class RecursiveDumpPageSourceCommand(BaseCommand):
         with gzip.GzipFile(outfile, "wb") as f:
             f.write(json.dumps(page_source).encode("utf-8"))
 
+class FinalizeCommand(BaseCommand):
+    """This command is automatically appended to the end of a CommandSequence
+    It's apperance means there won't be any more commands for this
+    visit_id
+    """
 
-def finalize(
-    visit_id: int, webdriver: WebDriver, extension_socket: clientsocket, sleep: int
-) -> None:
-    """ Informs the extension that a visit is done """
-    tab_restart_browser(webdriver)
-    # This doesn't immediately stop data saving from the current
-    # visit so we sleep briefly before unsetting the visit_id.
-    time.sleep(sleep)
-    msg = {"action": "Finalize", "visit_id": visit_id}
-    extension_socket.send(msg)
+    def __init__(self, sleep):
+        self.sleep = sleep
+
+    def __repr__(self):
+        return f"FinalizeCommand({self.sleep})"
+    
+    def execute(
+        self,
+        webdriver,
+        browser_settings,
+        browser_params,
+        manager_params,
+        extension_socket,
+    ):
+
+        """ Informs the extension that a visit is done """
+        tab_restart_browser(webdriver)
+        # This doesn't immediately stop data saving from the current
+        # visit so we sleep briefly before unsetting the visit_id.
+        time.sleep(self.sleep)
+        msg = {"action": "Finalize", "visit_id": self.visit_id}
+        extension_socket.send(msg)
 
 
 def initialize(visit_id: int, extension_socket: clientsocket) -> None:
